refactor: dto를 API 요청별로 분류
