+++ android/app/src/main/java/app/priceguard/data/dto/add/ProductAddRequest.kt
@@ -1,4 +1,4 @@
-package app.priceguard.data.dto
+package app.priceguard.data.dto.add
 
 import kotlinx.serialization.Serializable
