+++ android/app/src/main/java/app/priceguard/data/dto/add/ProductAddResponse.kt
@@ -1,9 +1,9 @@
-package app.priceguard.data.dto
+package app.priceguard.data.dto.add
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class PricePatchResponse(
+data class ProductAddResponse(
     val statusCode: Int,
     val message: String
 )
