+++ android/app/src/main/java/app/priceguard/data/dto/delete/ProductDeleteResponse.kt
@@ -1,4 +1,4 @@
-package app.priceguard.data.dto
+package app.priceguard.data.dto.delete
 
 import kotlinx.serialization.Serializable
