+++ android/app/src/main/java/app/priceguard/data/dto/patch/PricePatchRequest.kt
@@ -1,4 +1,4 @@
-package app.priceguard.data.dto
+package app.priceguard.data.dto.patch
 
 import kotlinx.serialization.Serializable
