+++ android/app/src/main/java/app/priceguard/data/dto/patch/PricePatchResponse.kt
@@ -1,9 +1,9 @@
-package app.priceguard.data.dto
+package app.priceguard.data.dto.patch
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ProductAddResponse(
+data class PricePatchResponse(
     val statusCode: Int,
     val message: String
 )
