+++ android/app/src/main/java/app/priceguard/data/dto/recommend/RecommendProductResponse.kt
@@ -1,4 +1,4 @@
-package app.priceguard.data.dto
+package app.priceguard.data.dto.recommend
 
 import kotlinx.serialization.Serializable
