+++ android/app/src/main/java/app/priceguard/data/dto/renew/RenewResponse.kt
@@ -1,4 +1,4 @@
-package app.priceguard.data.dto
+package app.priceguard.data.dto.renew
 
 import kotlinx.serialization.Serializable
