+++ android/app/src/main/java/app/priceguard/data/dto/signup/UserDataDTO.kt
@@ -1,4 +1,4 @@
-package app.priceguard.data.dto
+package app.priceguard.data.dto.signup
 
 import kotlinx.serialization.Serializable
