+++ android/app/src/main/java/app/priceguard/data/network/AuthAPI.kt
@@ -1,6 +1,6 @@
 package app.priceguard.data.network
 
-import app.priceguard.data.dto.RenewResponse
+import app.priceguard.data.dto.renew.RenewResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Header
