+++ android/app/src/main/java/app/priceguard/data/repository/TokenRepositoryImpl.kt
@@ -2,9 +2,9 @@ package app.priceguard.data.repository
 
 import android.util.Log
 import app.priceguard.data.datastore.TokenDataSource
-import app.priceguard.data.dto.RenewResult
-import app.priceguard.data.dto.UserDataDTO
-import app.priceguard.data.dto.UserDataResult
+import app.priceguard.data.dto.renew.RenewResult
+import app.priceguard.data.dto.signup.UserDataDTO
+import app.priceguard.data.dto.signup.UserDataResult
 import app.priceguard.data.network.APIResult
 import app.priceguard.data.network.AuthAPI
 import app.priceguard.data.network.getApiResult
