+++ android/app/src/main/java/app/priceguard/data/repository/UserRepositoryImpl.kt
@@ -1,11 +1,11 @@
 package app.priceguard.data.repository
 
-import app.priceguard.data.dto.LoginRequest
-import app.priceguard.data.dto.LoginResult
-import app.priceguard.data.dto.LoginState
-import app.priceguard.data.dto.SignupRequest
-import app.priceguard.data.dto.SignupResult
-import app.priceguard.data.dto.SignupState
+import app.priceguard.data.dto.login.LoginRequest
+import app.priceguard.data.dto.login.LoginResult
+import app.priceguard.data.dto.login.LoginState
+import app.priceguard.data.dto.signup.SignupRequest
+import app.priceguard.data.dto.signup.SignupResult
+import app.priceguard.data.dto.signup.SignupState
 import app.priceguard.data.network.APIResult
 import app.priceguard.data.network.UserAPI
 import app.priceguard.data.network.getApiResult
