+++ android/app/src/main/java/app/priceguard/ui/login/LoginViewModel.kt
@@ -3,7 +3,7 @@ package app.priceguard.ui.login
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.priceguard.data.dto.LoginState
+import app.priceguard.data.dto.login.LoginState
 import app.priceguard.data.repository.TokenRepository
 import app.priceguard.data.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
