+++ android/app/src/main/java/app/priceguard/ui/signup/SignupViewModel.kt
@@ -3,7 +3,7 @@ package app.priceguard.ui.signup
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.priceguard.data.dto.SignupState
+import app.priceguard.data.dto.signup.SignupState
 import app.priceguard.data.repository.TokenRepository
 import app.priceguard.data.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
