+++ android/app/src/main/java/app/priceguard/ui/splash/SplashScreenViewModel.kt
@@ -2,7 +2,7 @@ package app.priceguard.ui.splash
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.priceguard.data.dto.RenewResult
+import app.priceguard.data.dto.renew.RenewResult
 import app.priceguard.data.repository.TokenRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
